вывел функции в отдельный файл, массивы в отдельный, шаблоны страниц в отдельный, подключил все в index.php, написал функцию, которая обрезает теги у текста. обработал текст при выводе тегами.
