+++ data.php
@@ -1,0 +1,40 @@
+<?php
+$user_name = 'Алексей Онипко'; // укажите здесь ваше имя
+
+$popular_posts = [
+    [
+        'header' => 'Цитата',
+        'type' => 'post-quote',
+        'content' => 'Мы в жизни любим только раз, а после ищем лишь похожих',
+        'user-name' => 'Лариса',
+        'userpic' => 'userpic-larisa-small.jpg'
+    ],
+    [
+        'header' => 'Игра престолов',
+        'type' => 'post-text',
+        'content' => 'Не могу дождаться начала финального сезона своего любимого сериала! Не могу дождаться начала финального сезона своего любимого сериала! Не могу дождаться начала финального сезона своего любимого сериала! Не могу дождаться начала финального сезона своего любимого сериала! Не могу дождаться начала финального сезона своего любимого сериала! Не могу дождаться начала финального сезона своего любимого сериала!',
+        'user-name' => 'Владик',
+        'userpic' => 'userpic.jpg' 
+    ],
+    [
+        'header' => 'Наконец, обработал фотки!',
+        'type' => 'post-photo',
+        'content' => 'rock-medium.jpg',
+        'user-name' => 'Виктор',
+        'userpic' => 'userpic-mark.jpg' 
+    ],
+    [
+        'header' => 'Моя мечта',
+        'type' => 'post-photo',
+        'content' => 'coast-medium.jpg',
+        'user-name' => 'Лариса',
+        'userpic' => 'userpic-larisa-small.jpg' 
+    ],
+    [
+        'header' => 'Лучшие курсы',
+        'type' => 'post-link',
+        'content' => 'www.htmlacademy.ru',
+        'user-name' => 'Владик',
+        'userpic' => 'userpic.jpg' 
+    ]
+    ];
