+++ templates/layout.php
@@ -1,4 +1,6 @@
-
+<?php 
+$is_auth = rand(0, 1);
+?>
 <!DOCTYPE html>
 <html lang="ru">
 <head>
